changes PurchasePrice and DownPaymen tFrom Float To Int for financial table

+++ Models/Financial.cs
@@ -13,7 +13,8 @@ namespace LightningOffer.Models
         public Guid Financial_id { get; set; }
         public DateTime CreatedDate { get; set; }
         [Display(Name = "Purchase Price")]
-        public float Purchase_Price { get; set; }
+        
+        public int Purchase_Price { get; set; }
 
         [Display(Name = "Earnest Money Deposit")]
         public int EMD { get; set; }
